Fix 32-bit time_t compile errors in DLTv2 timestamp code
On platforms where time_t is 32-bit (e.g. armv7/musl, glibc arm with
default _FILE_OFFSET_BITS), several DLTv2 timestamp call sites trigger
-Werror=shift-count-overflow, -Werror=sign-conversion, or
-Werror=sign-compare, preventing compilation.

Fixes:

- ts.tv_sec >> 32 (9 sites in dlt-daemon.c, dlt_daemon_client.c,
  dlt_client.c, dlt_user.c, dlt_common.c): cast to uint64_t before
  shifting. On 32-bit time_t the shift is UB; on 64-bit it is a no-op.
  The top byte of the 40-bit seconds field is now correctly zero on
  32-bit platforms instead of carrying a stray duplicate of the low
  byte (caused by ARM's 5-bit shift-count mask).

- time_t tt = msg->storageheader->seconds (dlt_common.c:1084):
  explicit (time_t) cast to silence sign-conversion when time_t is
  signed 32-bit.

- st.st_size > UINT_MAX (dlt_user.c:5620): cast st.st_size to
  uintmax_t for the comparison, avoiding the signed/unsigned mismatch
  between off_t and unsigned int.

+++ src/daemon/dlt-daemon.c
@@ -2715,7 +2715,7 @@ int dlt_daemon_log_internal(DltDaemon *daemon, DltDaemonLocal *daemon_local,
 #else
         struct timespec ts;
         if(clock_gettime(CLOCK_REALTIME, &ts) == 0) {
-            msg.headerextrav2.seconds[0]=(uint8_t)((ts.tv_sec >> 32) & 0xFF);
+            msg.headerextrav2.seconds[0]=(uint8_t)(((uint64_t)ts.tv_sec >> 32) & 0xFF);
             msg.headerextrav2.seconds[1]=(uint8_t)((ts.tv_sec >> 24) & 0xFF);
             msg.headerextrav2.seconds[2]=(uint8_t)((ts.tv_sec >> 16) & 0xFF);
             msg.headerextrav2.seconds[3]=(uint8_t)((ts.tv_sec >> 8) & 0xFF);
@@ -2724,7 +2724,7 @@ int dlt_daemon_log_internal(DltDaemon *daemon, DltDaemonLocal *daemon_local,
                 msg.headerextrav2.nanoseconds = (uint32_t) ts.tv_nsec; /* value is long */
             }
         }else if (clock_gettime(CLOCK_MONOTONIC, &ts) == 0) {
-            msg.headerextrav2.seconds[0]=(uint8_t)((ts.tv_sec >> 32) & 0xFF);
+            msg.headerextrav2.seconds[0]=(uint8_t)(((uint64_t)ts.tv_sec >> 32) & 0xFF);
             msg.headerextrav2.seconds[1]=(uint8_t)((ts.tv_sec >> 24) & 0xFF);
             msg.headerextrav2.seconds[2]=(uint8_t)((ts.tv_sec >> 16) & 0xFF);
             msg.headerextrav2.seconds[3]=(uint8_t)((ts.tv_sec >> 8) & 0xFF);
